microservices/users_service/requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
Flask-Marshmallow
marshmallow-sqlalchemy
Flask-RESTx
prometheus_client
python-dotenv
python-consul
alembic
Flask-Migrate
pika
pytest
gunicorn